refactor: replace kodein with koin
- update gradle plugins
