+++ app/build.gradle
@@ -115,11 +115,11 @@ dependencies {
 
 
     // DI
-    implementation 'org.kodein.di:kodein-di-generic-jvm:5.2.0'
-    implementation 'org.kodein.di:kodein-di-framework-android-x:5.2.0'
+    implementation "org.koin:koin-android:2.1.5"
+    implementation 'org.koin:koin-android-viewmodel:2.1.5'
     // Firebase
-    implementation 'com.google.firebase:firebase-core:17.3.0'
-    implementation 'com.google.firebase:firebase-analytics:17.3.0'
+    implementation 'com.google.firebase:firebase-core:17.4.0'
+    implementation 'com.google.firebase:firebase-analytics:17.4.0'
     // Crashlytics
     implementation 'com.crashlytics.sdk.android:crashlytics:2.10.1'
     // Chrome custom tab
