+++ app/src/main/java/com/muhammadwahyudin/identitasku/di/AppModule.kt
@@ -1,0 +1,9 @@
+package com.muhammadwahyudin.identitasku.di
+
+import com.muhammadwahyudin.identitasku.ui.home.HomeViewModelImpl
+import org.koin.android.viewmodel.dsl.viewModel
+import org.koin.dsl.module
+
+val appModule = module {
+    viewModel { HomeViewModelImpl(get()) }
+}
